added calculations for flc

+++ src/main/java/com/leaps/dto/master/MedicalDetailsDto.java
@@ -15,7 +15,7 @@ public class MedicalDetailsDto {
 	private String prodName;
 	private String medicalCategory;
 	private String medicalCenter;
-	private Float mfRate;
+	private Double mfRate;
 	private String startDate;
 	private String endDate;
 }
